refactor(ios): split FileProviderExtension into focused services
Extract upload, download, mutation, and error-mapping logic out of the
monolithic FileProviderExtension into dedicated services so the extension
becomes a thin coordinator, easing testability and maintenance ahead of
move/trash/versioning work.
